pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-audio"
version = "0.1.0"
description = "Hybrid audio + text news story segmentation pipeline"
readme = "RREADME.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "News Audio Lab"}]
keywords = ["audio", "segmentation", "broadcast", "NLP", "radio"]
dependencies = [
    "typer>=0.12.3",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.1",
    "torchaudio>=2.2.0",
    "openai-whisper>=20231117",
    "pyannote.audio>=3.1.1",
    "pandas>=2.1.0",
    "tqdm>=4.66.0",
    "PyYAML>=6.0",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.1",
    "webrtcvad-wheels>=2.0.14",
    "azure-cognitiveservices-speech>=1.37.0",
    "python-dotenv>=1.0.1",
    "openai>=1.51.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[project.scripts]
codex-audio = "codex_audio.cli.app:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["codex_audio"]





